♻️ [Refactoring]: auto-layout.test.tsをKent C. Doddsのベストプラクティスに従って5つのファイルに分割し、テストケース名を日本語化
- auto-layout.test.tsを削除し、__tests__ディレクトリに以下の5ファイルに分割:
  - auto-layout.factory.test.ts - ファクトリメソッドのテスト
  - auto-layout.fromStyles.test.ts - スタイル変換のテスト
  - auto-layout.gap.test.ts - gap関連のテスト
  - auto-layout.padding.test.ts - padding関連のテスト
  - auto-layout.setters.test.ts - setterメソッドのテスト
- すべてのテストケース名を日本語に変更
- テストの機能とコードは変更なし
